Migrate enum arrays to the bitflags crate
`bitflags` is the "de-facto standard" crate when it comes to
representing bitfields in an easy manner, and providing the necessary
tooling around it.  It allows to easily construct and compare
combinations of flags without having to fall back to casts and slices of
individual flag bits.  In addition the type documents and guards
possible values, giving a more ergonomic user-experience overall.