property/control: Support non-linear enum->string conversion
On a driver where certain properties not necessarily start at zero, such
as an Intel card with the "scaling mode" property omitting the `None`
variant and only describing the values 1, 2, and 3, translating a value
of 3 ("Full aspect") runs into an index-out-of-bounds panic.  Similarly
many properties present an unsorted array of values.

Both problems are solved by initially checking whether the value is out
of bounds and otherwise comparing if that value interpreted as index
yields the same value from the values array - this is merely a
performance optimization and if not, the actual index for that value
needs to be linearly scanned before the translation to its string
representation can be made.