Add AsRef impl for DumbMapping
In softbuffer's DRM implementation, we need to have a way to get a [u8]
slice for our Deref implementation. Right now, we just create an entire
buffer of zeroes and read from that. As that is unidiomatic, this PR
adds an AsRef impl to DumbMapping to get a reference to the underlying
memory.

See:
https://github.com/rust-windowing/softbuffer/pull/135#discussion_r1288807990
Signed-off-by: John Nunley <dev@notgull.net>