Move the modifier from add_planar_framebuffer to PlanarBuffer
The modifier is a property of the buffer, so let’s put it where it
belongs.

This adds a validation that the flags correspond to the modifier.
Another option would be to remove the flag and instead rely on whether
the modifier is set or not to set it for the kernel, but I’m not sure
this actually improves anything.