control: Add a new HashMap for accessing properties by name
This is a different interface from what we were currently exposing to
users, instead of iterating on properties and calling get_property() on
each of them, do it once and store the results in a HashMap.

This lowers the amount of get_property() calls for clients which don’t
cache those results, and make the API a bit simpler.