drm-ffi: Call GETPROPERTY ioctl only once if no values or enums
When we call this ioctl first, we obtain the amount of values and enum
blobs available for this property, so that we can allocate the memory
and pass it in a second call.  But when there are zero values and zero
enum blobs, there is no need for any allocation or second call.

This further reduces the amount of DRM_IOCTL_MODE_GETPROPERTY ioctls
from 84 to 77 in the atomic_modeset example.