drm-sys: generate powerpc64 and riscv64 bindings
It's really important to have pregenerated bindings available for
platforms, because even though the use_bindgen option exists, it's not
possible to set it through Cargo when building something that
indirectly depends on drm-sys through a crate that doesn't set the
option.

Here, I've added bindings for the two platforms I'm easily able to
cross compile for that weren't already present.