Update CI to generate bindings on x86_64, and compare other targets
This pulls bindings from the Linux kernel repository, so we do not
depend on Ubuntu packages, or hacks to use packages from a different
arch. The parts that potentially differ are from sysroot headers.

CI compares each version of generated bindings against the x86_64
version, and fails if they do not all match.

This also adds `riscv64gc-unknown-linux-gnu` as a tested target.

Resolves https://github.com/Smithay/drm-rs/issues/165.