control: Add `Device::crtc_get/queue_sequence()` IOCTLs and event
These two IOCTLs are for querying the most recent sequence number (and
its "first pixel out" timestamp) as well as scheduling (queueing) an
event when a future sequence number (absolute or relative) is hit,
equally containing the "first pixel out" timestamp for that sequence.

As with `wait_vblank()`, `crtc_queue_sequence()` also returns the
absolute sequence number that the event ended up latching on.  This
is useful when specifying a relative index as well as when the driver
`max()`es an absolute target to the most recently triggered sequence.
This request will still result in an event even when that sequence
number has already been hit, but doesn't allow you to queue older
frames.