Make `control::property::Info::value_type` return `&` instead of clone
It seems odd for this to return an owned value, when `::name()` doesn't.
And it is useful to get a reference to the `ValueType`, so
`ValueType::convert_value` can be used and just use the lifetime of the
`Info` instead of the cloned `ValueType.`

I assume there's no reason not to do this, since callers can still clone
as needed.